Add a CORS header allowing browser-based clients to consume the API.
The MaveDB API is publicly accessible and can be consumed by any client. However, it is currently consumed mainly by server-side clients such as mavevis. Browser-based clients, including any MaveDB client application that is not served by the Django application, generally require CORS headers granting explicit permission to consume the API. To allow this, we add a simple HTTP response header granting access to clients from any domain:
  Access-Control-Allow-Origin="*"

The header is added using Django middleware, with support for Django 1 (though a draft of a version for later Django versions is included and commented out). For current Django versions, there are also Django middleware packages available for this purpose, which provide configuration-driven CORS headers and are useful in case one wants to limit access to specific clients.